Welcome automation loversgit add .! You've come to the right place, because this tutorial is all about Ansible! And boy, it's just a lot of fun. Imagine if we could take control of an army of tiny, but mighty robots... simply by writing a YAML file. Think of the servers we could launch! The infrastructure we could build! The code we could deploy! The laundry we could wash! Well, actually, these Ansible robots are sorta virtual robots, so yes to automating things like server setup... but probably not to automating your laundry. However, I do hope one of you proves me wrong!
The App & Our Mission!

The first version of our app is done actually! But there are two problems. First, we'd love to have a way to easily boot up development servers complete with PHP, a web server and anything else we need to get our app working. And second, it would be bananas if we could have an automated way to deploy code to Amazon EC2 servers. Well, that's exactly what we're going to do.

Our project is a Symfony application... but if you're not used to Symfony, that's no problem. From a server-perspective, the app requires a few interesting things:

111 lines src/AppBundle/Controller/DefaultController.php
... lines 1 - 9
class DefaultController extends Controller{
... lines 12 - 14
    public function indexAction(){
        $ videos = $ this->getVideoRepository()
            ->findAll();
        $ tags = $ this->getUniqueOrderedTags($ videos);
        // Redis cache
        try {
            if ($ this->getRedisClient()->exists('total_video_uploads_count')) {
                $ totalVideoUploadsCount = $ this->getRedisClient()->get('total_video_uploads_count');
            } else {
                $ totalVideoUploadsCount = $ this->countTotalVideoUploads();
                $ this->getRedisClient()->set('total_video_uploads_count', $ totalVideoUploadsCount, 'ex', 60); // 60s
            }
... lines 29 - 35
        } catch (ConnectionException $ e) {
            $ totalVideoUploadsCount = $ this->countTotalVideoUploads();
            $ totalVideoViewsCount = $ this->countTotalVideoViews();
        }
        return $ this->render('default/index.html.twig', [
            'videos' => $ videos,
            'tags' => $ tags,
            'totalVideoUploadsCount' => $ totalVideoUploadsCount,
            'totalVideoViewsCount' => $ totalVideoViewsCount,
        ]);
    }
... lines 48 - 109
}

First, it needs a database connection to load the video info. And second, it uses Redis to cache a few things. So when we boot up our servers, that stuff needs to be there.
Introducing Ansible

There's a good chance you've setup a server before. I know I've setup a ton, and it's always the same, manual, error-prone, confusing process: SSH into the server, manually run a bunch of commands, have errors, Google things, run more commands, and edit a bunch of files.

Ansible... kinda does the same thing. When you execute something with Ansible, it ultimately SSH's onto the server and runs some commands. Ansible is said to be agentless, which just means that you don't need to install anything on the target server. As long as you can SSH to a server, you can unleash your Ansible robot army onto it. That's pretty cool.

But Ansible is even more interesting than that. When you execute an Ansible task - that's what they're called - it is idempotent... well, usually it is. Idempotency is an obscure - but cool - word to mean that Ansible tasks don't just dumbly run a command. What they really do is guarantee that the server finishes in a specific state. For example, if we tell Ansible to create a directory, it doesn't necessarily mean that it will run a mkdir command. Instead, it means that Ansible will just make sure that the directory exists - only creating it if necessary.

This idea of guranteeing a state - like 'this directory must exist' - is much more powerful than randomly running commands over SSH. These tasks also send back JSON info about what happened, which we can use to tweak and influence what happens next.