Ansible works by connecting over SSH and then running commands. Well, technically, that's not 100% true: you can actually configure Ansible to connect to your server in different ways, though you'll almost always use SSH. The most common exception is when you're working on your local machine - you don't need to connect via SSH at all!
To tell Ansible that this is a local connection, in your hosts.ini file, after the IP address, add ansible_connection=local

There's also a docker connection type if you're getting nerdy with Docker.

Try that ping again!

ansible 127.0.0.1 -m ping -i ansible/hosts.ini

Got it!

This little change is actually really important. By saying ansible_connection=local, we are setting a variable inside of Ansible. And as we build out more complex Ansible configuration, this idea of setting and using variables will become more important. As you'll see, you can set more variables for each host, which will let us change behavior on a host-by-host basis.

In this case, ansible_connection is a built-in variable that Ansible uses when it connects. We're simply changing it first.