Let's start filling in the playbook. At the top, every YAML file in Ansible starts with three dashes (---) t's not really important and doesn't meaning anything... it's just a YAML standard. Below, add host: all. Then, indent two spaces, add tasks, indent again, and add ping: ~. If you're new to YAML, you do need to be careful: the spacing and indentation are important. Try it:
ansible-playbook ansible/playbook.yml -i ansible/hosts.ini

Sweet! It runs the ping module against all hosts: our local machine and the VM.

Ok, back to the terminology we saw in the error message. A playbook contains plays. In this case we have just one play that will run on all hosts. Later, if we added another host line with its own tasks below it, that would be a second play. So, a playbook contains plays, a play contains tasks and each task executes a module. In this case we're executing the ping module.