Our app is written in PHP... so... we should probably get that installed. Copy the git block, paste it, and change it to php5-cli Run that playbook!
ansible-playbook ansible/playbook.yml -i ansible/hosts.ini

And yea... I know PHP 5 is ancient. But our Ubuntu distribution doesn't support version 7. But don't worry, we're going to upgrade to 7 in a minute by using a custom repository. It's going to be awesome - so stay with me.

While the playbook is running, change over to your terminal tab and make sure you're still SSH'ed into the VM. Of course, if we try php -v right now... it doesn't work. But as soon as Ansible finishes, try it again:

php -v

Yea! Version 5.5.9. Now, let's kill this ancient version of PHP and go to 7.