Using a Custom apt Repository
Here's the deal: if you research how to install PHP 7 on this version of Ubuntu, you'll learn about a third-party repository called ppa:ondrej/php. If we can add this to our apt sources - usually done by running a few commands - we'll be in business.

And of course... there's a module for that! It's not apt, it's apt_repository. It doesn't have any requirements, and the options look pretty easy - just set repository to the one we want to use.

Let's do it! In the playbook, above the PHP task, add a new one: Add PHP 7 Personal Package Archive repository. Use the apt_repository module and set repository to: ppa:ondrej/php.

And now, we can install php7.1-cli

Run it!

ansible-playbook ansible/playbook.yml -i ansible/hosts.ini

Love it... no errors, and 2 tasks changed. Over in the VM, try:

php -v

Oh, sweet PHP 7.1 goodness.