The git Module
While we're waiting, let's move to the next step: cloning our code via Git. Unfortunately, Ansible does not have a Git module to help us... bah! Just kidding, it totally does! Search for the 'Ansible git module' to find it.

The module does have Git as a requirement - but we already installed that. So our job is pretty simple: pass it the repo we want to clone, and the dest-ination we want to clone to.

Do it! Go to http://github.com/symfony/symfony-standard. To start, we'll pull down the Symfony Standard Edition instead of our MooTube code. But that's just temporary. I'll click the 'Clone' button and copy the URL.

Now, add a task: Checkout Git repository. We do not need become: true because we own the destination directory. Go straight to git, then repo set to the URL we just copied. For dest, put /var/www/project. Add force: yes: that'll discard uncommitted changes if there are any

Head back to your terminal. Sweet! The directory was created! In the VM, the /var/www/project directory is empty.