With our project cloned, the next step is obvious: use Composer to install our dependencies! Actually, we used the Composer module earlier from the command line. Google again for 'Ansible Modules' and find the 'All Modules' page. I'll find the composer module and open that in a new tab.
This module is really easy... except for one problem. Under Requirements, it says that Composer already needs to be installed. We have not done that yet... and, unfortunately, it can't be installed with apt-get.
Installing Composer Programmatically?

So how do you install it? Check out https://getcomposer.org and click 'Download'.

Normally, we just paste these lines into our terminal and celebrate! But... there's this problematic fine print at the bottom:

    Do not redistribute the install code. It will change for every version of the install.

Huh. Composer includes a bit of built-in security: a sha hash to make sure that the installer hasn't been tampered with. If we tried to use these 4 commands in Ansible, it would work... for awhile. But next time the installer is updated, and that sha changed... it would stop working.

What to do? Check out that how to install Composer programmatically link(https://getcomposer.org/doc/faqs/how-to-install-composer-programmatically.md). Eureka: a shell script that will safely download the latest version of Composer. The end result is a composer.phar file wherever we run this script from.

Installing Composer with script

Our mission is clear: somehow, execute this shell script via Ansible. How can we execute a script on a host? Why, with... the script module of course!
Neato! We just point it at a local script, and it takes care of the rest. Go copy the script and, in our ansible directory, create a new scripts directory and a new file called install_composer.sh. Paste the code there.

Back in the playbook, at the bottom, create a new task: Download Composer. Use the script module. Then, the easiest way to use this is to literally put the script filename on the same line as the module name: script: scripts/install_composer.sh

Actually, every module can be used with a one-line syntax like this... but since line breaks are pretty cheap these days, I usually organize things a bit more.

Thanks to this task, we'll have a new composer.phar file in our home directory, which is where this task - well, all tasks - are running. But that's not enough: we need to move this to /usr/local/bin/composer.