Nginx is still using its generic, boring, but very polite default HTML page:
    You're welcome for using Nginx!

It's time to point Nginx at our code! Google for 'Symfony Nginx' to find a documentation page about Configuring a Web Server.

Scroll down to the Nginx section: it has a great block of Nginx config that'll work with our app. Here's the trick: we somehow need to put this configuration up onto the server and customize it. How can we do that?

Ansible does have a module for copying files from your local machine to the remote host. But in this case, there's a different module that's even more perfect: the template module. It says:

    Templates a file out to a remote server

In normal-person English, this means that it executes a file through Jinja - allowing us to print dynamic variables - and then copies it up to the remote server. I love this module.

Create a templates/ directory and inside, a symfony.conf file. Paste in the raw Nginx configuration.

Now, a few things need to change in here. Like, server_name should be mootube.l and root should point to the correct directory. Actually, root is already correct by chance... but we can do better!