I just ran the playbook with the deploy tag:
ansible-playbook ansible/playbook.yml -i ansible/hosts.ini -t deploy

Notice that several tasks say 'Changed'... but that's a lie! The first two are related to Composer - we'll talk about those later. Right now, I want to focus on the last 4: fixing the directory permissions and the 3 bin/console commands:

195 lines ansible/playbook.yml
---
- hosts: vb
... lines 3 - 10
  tasks:
... lines 12 - 156
    - name: Fix var directory permissions
      file:
        path: '{{ symfony_var_dir }}'
        state: directory
        mode: 0777
        recurse: yes
      tags:
        - permissions
        - deploy
    # Symfony console commands
    - name: Create DB if not exists
      command: '{{ symfony_console_path }} doctrine:database:create --if-not-exists'
      tags:
        - deploy
    - name: Execute migrations
      command: '{{ symfony_console_path }} doctrine:migrations:migrate --no-interaction'
      tags:
        - deploy
    - name: Load data fixtures
      command: '{{ symfony_console_path }} hautelook_alice:doctrine:fixtures:load --no-interaction'
      tags:
        - deploy
... lines 182 - 195

Changed and Idempotency

But first... why do we care? I mean, sure, it says 'Changed' when nothing really changed... but who cares? First, let me give you a fuzzy, philosophical reason. Tasks are meant to be idempotent... which is a hipster tech word to mean that it should be safe to run a task over and over again without any side effects.

And in reality, our tasks are idempotent. If we run this 'Fix var directory permissions' task over and over and over again... that's fine! Nothing weird will happen. It's simply that the tasks are reporting that something is changing each time... when really... it's not!

I know, I know... this seems like such a silly detail. But soon, we're going to start making decision in our playbook based on whether or not a task reports as 'changed'.

Actually, this is already happening:

195 lines ansible/playbook.yml
---
- hosts: vb
... lines 3 - 10
  tasks:
... lines 12 - 182
  handlers:
    - name: Restart Nginx
      become: true
      service:
        name: nginx
        state: restarted
... lines 189 - 195

The 'Restart Nginx' handler is only called when this task changes:

195 lines ansible/playbook.yml
---
- hosts: vb
... lines 3 - 10
  tasks:
... lines 12 - 56
    - name: Enable Symfony config template from Nginx available sites
      become: true
      file:
        src: '/etc/nginx/sites-available/{{ server_name }}.conf'
        dest: '/etc/nginx/sites-enabled/{{ server_name }}.conf'
        state: link
      notify: Restart Nginx
... lines 64 - 195

So, as a best practice - as much as we can - we want our tasks to correctly report whether or not they changed.